email triage agent/requirements.txt
# Core
langchain
langchain-core
python-dotenv
pydantic

# LLM backends (install the ones you use)
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-google-vertexai
